fix: Fix data batching by adding wraparound
Previously, words at the end were discarded due to invalid sequence
length. When the current loop pointer added with the sequence length
exceeds the total length of words, remaining words can be calculated and
these number of words from the start of the list are appended to the
remaining words at the end.